resoved comments
