+++ src/stepFunctions/commands/visualizeStateMachine/aslVisualizationManager.ts
@@ -56,15 +56,7 @@ export class AslVisualizationManager extends AbstractAslVisualizationManager {
 
             return newVisualization.getPanel()
         } catch (err) {
-            vscode.window.showInformationMessage(
-                localize(
-                    'AWS.stepfunctions.visualisation.errors.rendering',
-                    'There was an error rendering State Machine Graph, check logs for details.'
-                )
-            )
-
-            logger.debug('Unable to setup webview panel.')
-            logger.error(err as Error)
+            this.handleErr(err, logger)
         }
 
         return
