+++ src/test/cdk/tree/treeInspector.test.ts
@@ -115,78 +115,4 @@ describe('TreeInspector', function () {
 
         assert.strictEqual(props, undefined)
     })
-
-
-    it('returns true when tree node contains a node with id === "Resource" and type === "StateMachine"', async function () {
-        const construct: ConstructTreeEntity = {
-            id: 'StateMachine',
-            path: 'aws-stepfunctions-integ/StateMachine',
-            children: { 'Resource': {
-                            id: 'Resource',
-                            path: 'aws-stepfunctions-integ/StateMachine/Resource',
-                            attributes: {
-                                "aws:cdk:cloudformation:type": 'AWS::StepFunctions::StateMachine'
-                            }
-                        }
-                      }
-
-        }
-
-        assert.ok(treeInspector.isStateMachine(construct))
-    })
-
-    it('returns true when tree node contains a node with id !== "Resource" and type === "StateMachine"', async function () {
-        const construct: ConstructTreeEntity = {
-            id: 'StateMachine',
-            path: 'aws-stepfunctions-integ/StateMachine',
-            children: { 'Other': {
-                            id: 'Other',
-                            path: 'aws-stepfunctions-integ/StateMachine/Resource',
-                            attributes: {
-                                "aws:cdk:cloudformation:type": 'AWS::StepFunctions::StateMachine'
-                            }
-                        }
-                      }
-
-        }
-
-        assert.strictEqual(treeInspector.isStateMachine(construct),false)
-    })
-
-    it('returns false when tree node contains a node with id !== "Resource" and type !== "StateMachine"', async function () {
-
-        const construct: ConstructTreeEntity = {
-            id: 'StateMachine',
-            path: 'aws-stepfunctions-integ/LambdaFunction',
-            children: { 'Other': {
-                            id: 'Other',
-                            path: 'aws-stepfunctions-integ/LambdaFunction/Resource',
-                            attributes: {
-                                "aws:cdk:cloudformation:type": 'AWS::StepFunctions::LambdaFunction'
-                            }
-                        }
-                      }
-
-        }
-        assert.strictEqual(treeInspector.isStateMachine(construct),false)
-    })
-
-
-    it('returns false when tree node contains a node with id === "Resource" and type !== "StateMachine"', async function () {
-
-        const construct: ConstructTreeEntity = {
-            id: 'StateMachine',
-            path: 'aws-stepfunctions-integ/LambdaFunction',
-            children: { 'Resource': {
-                            id: 'Resource',
-                            path: 'aws-stepfunctions-integ/LambdaFunction/Resource',
-                            attributes: {
-                                "aws:cdk:cloudformation:type": 'AWS::StepFunctions::LambdaFunction'
-                            }
-                        }
-                      }
-
-        }
-        assert.strictEqual(treeInspector.isStateMachine(construct),false)
-    })
 })
